Added styling for the main page

+++ CMCS_WebApplication/Models/Claim.cs
@@ -9,10 +9,10 @@ namespace CMCS_WebApplication.Models
         public string RowKey { get; set; }
         public int Lecturer_ID { get; set; }
         public int Claim_ID { get; set; }
-        public decimal Hours_Worked { get; set; }
-        public decimal Hourly_Rate { get; set; }
+        public int Hours_Worked { get; set; }
+        public int Hourly_Rate { get; set; }
         public string Claim_Status { get; set; }
-        public decimal Final_Pay { get; set; }
+        public int Final_Pay { get; set; }
         public DateTimeOffset? Timestamp { get; set; }
         public ETag ETag { get; set; }
     }
